pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["providers"]

[project]
name = "explore-osworld"
version = "0.1.0"
description = "My research and experimentation on OSWorld"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "daytona>=0.112.1",
    "docker>=7.1.0",
    "filelock>=3.20.0",
    "pillow>=12.0.0",
    "psutil>=7.1.2",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
